requirements.txt
easydict==1.13
librosa==0.9.2
matplotlib==3.5.1
nnAudio==0.3.3
numpy==1.24.1
pandas==1.4.1
pytorch-lightning==1.6.0
scikit-learn==1.0.2
SoundFile==0.10.3.post1
tqdm==4.66.5
transformers==4.15.0
